pyproject.toml
[tool.poetry]
name = "otakumandu"
version = "0.1.0"
description = ""
authors = ["CaffeineDuck <samrid.pandit@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = "^1.7.3"
tortoise-orm = {extras = ["asyncpg"], version = "^0.17.6"}
pydantic = {extras = ["dotenv"], version = "^1.8.2"}
aerich = "^0.5.5"
jishaku = "^2.2.0"

[tool.poetry.dev-dependencies]
black = "^21.7b0"
isort = "^5.9.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
